+++ gfx_text/src/text.h
@@ -39,7 +39,7 @@ extern SDL_Rect tempTextureRect;
 extern TTF_Font *currentFont;
 
 
-char *hexToString(short hex);
+char *hexToString(int hex);
 SDL_Texture* loadText(SDL_Renderer *renderer, const char *text, int size, SDL_Color color);
 
 void addToTextArrays(struct textData *textData, short *value);
